+++ bolt/common/base/SimdUtil.h
@@ -465,12 +465,12 @@ inline T* addBytes(T* pointer, int32_t bytes) {
 // 'memcpy' implementation that copies at maximum width and unrolls
 // when 'bytes' is constant.
 template <typename A = xsimd::default_arch>
-inline void memcpy(void* to, const void* from, int32_t bytes, const A& = {});
+inline void memcpy(void* to, const void* from, size_t bytes, const A& = {});
 
 // memset implementation that writes at maximum width and unrolls for
 // constant values of 'bytes'.
 template <typename A = xsimd::default_arch>
-void memset(void* to, char data, int32_t bytes, const A& = {});
+void memset(void* to, char data, size_t bytes, const A& = {});
 
 // Calls a different instantiation of a template function according to
 // 'numBytes'.
